Docusaurus init and jwt auth implementation first steps

+++ backend/db_schema/letterboxd.sql
@@ -5,14 +5,14 @@ DROP TABLE IF EXISTS users;
 
 -- Users table
 CREATE TABLE users (
-    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     username VARCHAR(50) NOT NULL UNIQUE
 );
 
 -- Follows table 
 CREATE TABLE follows (
-    follower_id BIGINT NOT NULL,
-    following_id BIGINT NOT NULL,
+    follower_id INT NOT NULL,
+    following_id INT NOT NULL,
     followed_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (follower_id, following_id),
     FOREIGN KEY (follower_id) REFERENCES users(id) ON DELETE CASCADE,
@@ -22,7 +22,7 @@ CREATE TABLE follows (
 
 -- Film table 
 CREATE TABLE film (
-    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     title VARCHAR(255) NOT NULL,
     year INT CHECK (year > 1800 AND year <= EXTRACT(YEAR FROM CURRENT_DATE)),
     director VARCHAR(255),
@@ -34,8 +34,8 @@ CREATE TABLE film (
 
 -- Film Ratings table
 CREATE TABLE film_ratings (
-    user_id BIGINT NOT NULL,
-    film_id BIGINT NOT NULL,
+    user_id INT NOT NULL,
+    film_id INT NOT NULL,
     rating DECIMAL(2, 1) CHECK (rating >= 0 AND rating <= 10),
     rated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (user_id, film_id),
